creating required folders for unity package, Runtime, Editor and Samples~ moving files into appropriate directories, deleting meta files
