--- GameFuse.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3eba033ef79524fbf8acf740b53e0ec9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
