--- GameFuseExample.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9bdaca315d7fc47dfade65cd2b061d00
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
