--- GameFuseExample.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 65406a14153894b4bb107385db3bbef0
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
