--- GameFuseInitializer.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a36bca416db09434a9fa8421564d4fc1
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
