--- GameFuseLeaderboardEntry.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c96465766983a4bef90cd201c366516f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
