--- GameFuseStoreItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 80be5cb6ebd9c4194b629159c08af75f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
