--- GameFuseUser.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c2242185fcdf7431fb45e143efb84794
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
