--- GameFuseUtilities.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1dab052198ace4591a31431358539ea3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
