--- JSONObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0f239d45c23db4e148fade0194656a7e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
